Make really single-process runtime that handles fork
This scheme is quite fragile, but works.

The idea is to remove blocking operation, and blocking pool is not
created at all.  For reqwest, it is DNS resovling; we use trust-dns
crate.

Another, non-related problem is the connection pool, even if we access
new host.  We set ClientBuilder::pool_idle_timeout to zero seconds to
avoid it.