Revert "Handle chain of forward pointers"
This reverts commit 0a051c07bbc455c71cc83cba2217ac5ec483648e.

Chain of forward pointers is never created, as the older generation is
always larger than the younger one, and if older generation has no
enough memory for the whole younger generation, next generation is
used.

When oldest generation is used, the GC may overflow indeed, but it is
a fatal failure not creating chained forward pointers.